Updated the ADIN1100 Driver with version adin1100_v1.0.0.

+++ FW_Package/10BASE-T1L-MC_SourceCode/bsp/boardsupport.c
@@ -262,7 +262,7 @@ uint32_t UartInit(void)
 
 uint32_t BSP_InitSystem(void)
 {
-    uint32_t error = SUCCESS;
+    uint8_t error = SUCCESS;
 
     do
     {
